+++ svelte/src/components/wrapper.tsx
@@ -1,14 +1,15 @@
 import { API } from '@storybook/api';
 
 import { STORY_CHANGED, STORY_RENDERED } from '@storybook/core-events';
-import React, { useState } from 'react';
+import React, { useRef, useState } from 'react';
 
 // Component
 import InteractionPanel from './panel';
 // Helper
-import { DEFINE_ACTORS, WS_PORT } from '../helpers/constants';
+import { DEFINE_ACTORS, GLOBAL_ACTORS_IDENTIFIER, WS_PORT } from '../helpers/constants';
 // Type
 import type { Actors, Framework, ObjectExpression } from '../../types';
+import { getLocalStorage } from '../helpers';
 
 // WebSocket client
 const ws = new WebSocket(`ws://localhost:${WS_PORT}`);
@@ -20,10 +21,17 @@ export default ({ api, active }: { api: API, active: boolean }) => {
   const [ componentName, setComponentName ] = useState<string>('');
   const [ extension, setExtension ] = useState<string>('');
   const [ props, setProps ] = useState<ObjectExpression['properties']>([]);
+
+  const globalActorsDefinition = useRef<Actors>(undefined);
+
   // Listen for future story selections
   api.on(STORY_RENDERED, () => {
     const currentStory = api.getCurrentStoryData();
-    setActors((currentStory?.parameters as any)?.[DEFINE_ACTORS]);
+    // Prioritise actors defined specifically in the stories file
+    if (!globalActorsDefinition.current) {
+      globalActorsDefinition.current = getLocalStorage.getItem(GLOBAL_ACTORS_IDENTIFIER);
+    }
+    setActors((currentStory?.parameters as any)?.[DEFINE_ACTORS] || globalActorsDefinition.current);
     setRenderedState(true);
     setFramework((currentStory?.parameters as any).framework);
 
@@ -38,6 +46,12 @@ export default ({ api, active }: { api: API, active: boolean }) => {
   });
   ws.onmessage = (e) => {
     const response = JSON.parse(e.data);
+    if (response.key === GLOBAL_ACTORS_IDENTIFIER && !globalActorsDefinition.current) {
+      // Save this to localStorage for future use as we don't want to rely on the preset side all the time
+      getLocalStorage.setItem(response.key, response.actors);
+      return;
+    }
+
     setComponentName(response.fileName);
     setExtension(response.fileExtension);
     setProps(response.exportDeclarations);